requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
nest-asyncio==1.5.8
arxiv==1.4.8
openai>=1.0.0
eventlet==0.33.3
mcp
